requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
Flask-CORS==4.0.0
Flask-Mail==0.9.1
google-generativeai==0.7.2
bcrypt==4.0.1
PyJWT==2.8.0
cryptography==41.0.7
python-dotenv==1.0.0
stripe==9.12.0
gunicorn==21.2.0
pymongo==4.6.0
certifi==2024.2.2
